pyproject.toml
[project]
name = "tokenizer-visualizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hf-xet>=1.2.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "tiktoken>=0.12.0",
    "tokenizers>=0.22.1",
    "transformers>=4.57.1",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
